made texture lookups uint8_t based, which packs memory a lot tighter in various structures. it implies 255 max textures which might hurt some wads, but the commercial games seem ok.

+++ doomstat.h
@@ -236,7 +236,7 @@ extern  int8_t             bodyqueslot;
 // Needed to store the number of the dummy sky flat.
 // Used for rendering,
 //  as well as tracking projectiles etc.
-extern int16_t		skyflatnum;
+extern uint8_t		skyflatnum;
